requirements.txt
fastapi
uvicorn
scrapy
psycopg2-binary
